Some changes that speed up the candidate selections for alpha and beta peptide. Instead of getting a sorted list of all possible alpha peptides or all possible peptide pairs only the top n will be sorted.
Discarded decoys are no longer replaced by random peptides but if 
possible by permutations of the discarded peptides. If no permutation
can be found, that has not been registered yet, then it will really 
discard the peptide.

This should work reasonable well large peptides but for small peptides 
the chance to find something will depend very much on the size of the
searched database. E.g. in a proteome scale search the chance to find a 
new permutation of a 3 amino-acid peptide will be close to nonexistent.